microservices_s/service_upload/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-multipart
pydantic==2.10.6
cassandra-driver
minio
httpx==0.28.1
python-dotenv==1.0.1
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
ecdsa==0.19.1
h11==0.14.0
httpcore==1.0.7
idna==3.10
jwt==1.3.1
pyasn1==0.4.8
pycparser==2.22
pydantic_core==2.27.2
python-jose==3.4.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
pydantic
fastapi
python-dotenv
cassandra-driver
minio
